// src/app/model/CardModel.ts
export class Card {
    country: string;
    population: number;
    year: number;
    maleUrban: number;
    maleRural: number;
    maleTotal: number;
    femaleUrban: number;
    femaleRural: number;
    femaleTotal: number;
    bothUrban: number;
    bothRural: number;
    bothTotal: number;
    sourceYear: number;
}